chore: rename addons to enhancers
This is to be inline with eslint-config naming.
