+++ .eslintrc.json
@@ -1,8 +1,10 @@
 {
     "root": true,
+    "env": {
+        "node": true
+    },
     "extends":[
-        "@moxy/eslint-config/es9",
-        "@moxy/eslint-config/addons/node",
-        "@moxy/eslint-config/addons/jest"
+        "@moxy/eslint-config-base",
+        "@moxy/eslint-config-jest"
     ]
 }
