+++ lib/compose.js
@@ -1,11 +1,11 @@
 'use strict';
 
-module.exports = (listOfAddons = []) => {
-    if (!Array.isArray(listOfAddons)) {
+module.exports = (listOfEnhancers = []) => {
+    if (!Array.isArray(listOfEnhancers)) {
         throw new Error('Received argument is not an array.');
     }
 
-    return listOfAddons.reduce((acc, addon) => {
+    return listOfEnhancers.reduce((acc, addon) => {
         if (typeof addon !== 'function') {
             throw new Error('Found item in argument list that is not a function, all objects to compose should be functions!');
         }
