+++ lib/enhancers/index.js
