+++ lib/enhancers/with-enzyme/index.js
