+++ lib/enhancers/with-enzyme/index.test.js
