+++ lib/enhancers/with-enzyme/setup-adapter.js
