+++ lib/enhancers/with-enzyme/setup.js
