+++ lib/index.js
@@ -2,6 +2,6 @@
 
 const compose = require('./compose');
 const baseConfig = require('./base-config');
-const addons = require('./addons');
+const enhancers = require('./enhancers');
 
-module.exports = { compose, baseConfig, ...addons };
+module.exports = { compose, baseConfig, ...enhancers };
