+++ package.json
@@ -32,7 +32,8 @@
     "@babel/preset-env": "^7.7.1",
     "@commitlint/cli": "^8.2.0",
     "@commitlint/config-conventional": "^8.2.0",
-    "@moxy/eslint-config": "^10.1.1",
+    "@moxy/eslint-config-base": "^11.0.0",
+    "@moxy/eslint-config-jest": "^11.0.3",
     "eslint": "^6.5.1",
     "husky": "^4.2.0",
     "jest": "^25.1.0",
